# docs/swagger.yaml
basePath: /api/v2
definitions:
  apperrors.ErrorResponse:
    properties:
      error_response:
        type: string
    type: object
  doc_structs.AllWorkspacesResponse:
    properties:
      workspaces:
        $ref: '#/definitions/dto.AllWorkspaces'
    type: object
  doc_structs.AvatarUploadResponse:
    properties:
      avatar_url:
        $ref: '#/definitions/dto.UrlObj'
    type: object
...
  dto.AllWorkspaces:
    properties:
      guestWorkspaces:
        items:
          $ref: '#/definitions/dto.UserGuestWorkspaceInfo'
        type: array
      yourWorkspaces:
        items:
          $ref: '#/definitions/dto.UserOwnedWorkspaceInfo'
        type: array
    type: object
...
  dto.AvatarChangeInfo:
    properties:
      avatar:
        items:
          type: integer
        type: array
      user_id:
        type: integer
    type: object
...
  dto.NewWorkspaceInfo:
    properties:
      description:
        type: string
      name:
        type: string
      owner_id:
        type: integer
    type: object
  dto.PasswordChangeInfo:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.UpdatedWorkspaceInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.UrlObj:
    properties:
      url:
        type: string
    type: object
  dto.UserGuestWorkspaceInfo:
    properties:
      boards:
        items:
          $ref: '#/definitions/dto.WorkspaceBoardInfo'
        type: array
      workspace_id:
        type: integer
      workspace_name:
        type: string
      workspace_owner:
        $ref: '#/definitions/dto.UserOwnerInfo'
    type: object
...
  dto.UserOwnedWorkspaceInfo:
    properties:
      boards:
        items:
          $ref: '#/definitions/dto.WorkspaceBoardInfo'
        type: array
      workspace_id:
        type: integer
      workspace_name:
        type: string
    type: object
  dto.UserOwnerInfo:
    properties:
      owner_email:
        type: string
      owner_id:
        type: integer
      owner_name:
        type: string
      owner_surname:
        type: string
    type: object
  dto.UserProfileInfo:
    properties:
      description:
        type: string
      name:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.WorkspaceBoardInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.WorkspaceID:
    properties:
      workspace_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: kdanil01@mail.ru
    name: Капитанов Даниил
    url: https://vk.com/poophead27
  description: Лучшее и единственно приложение, имитирующее Trello.
  license:
    name: None
    url: None
  title: LA TABULA API
  version: "2.0"
paths:
...
  /user/edit/:
    post:
      consumes:
      - application/json
      description: В ответ ничего не шлёт
      parameters:
      - description: id пользователя, имя, фамилия, описание пользователя
        in: body
        name: newProfileInfo
        required: true
        schema:
          $ref: '#/definitions/dto.UserProfileInfo'
      produces:
      - application/json
      responses:
        "200":
          description: no content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Поменять данные профиля
      tags:
      - auth
  /user/edit/change_avatar/:
    post:
      consumes:
      - application/json
      description: В ответ шлёт ссылку на файл
      parameters:
      - description: id пользователя, изображение
        in: body
        name: avatarChangeInfo
        required: true
        schema:
          $ref: '#/definitions/dto.AvatarChangeInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Объект пользователя
          schema:
            $ref: '#/definitions/doc_structs.AvatarUploadResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Поменять аватарку
      tags:
      - auth
  /user/edit/change_password/:
    post:
      consumes:
      - application/json
      description: Получает старый и новый пароли, а также id пользователя
      parameters:
      - description: id, старый и новый пароли пользователя
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/dto.PasswordChangeInfo'
      produces:
      - application/json
      responses:
        "200":
          description: no content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Поменять пароль
      tags:
      - user
  /user/workspaces:
    get:
      consumes:
      - application/json
      description: Выводит и созданные им рабочие пространства и те, в которых он
        гость. Работает только для авторизированного пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь и его рабочие пространства
          schema:
            $ref: '#/definitions/doc_structs.AllWorkspacesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Вывести все рабочие пространства и доски текущего пользователя
      tags:
      - workspaces
  /workspace/create/:
    post:
      consumes:
      - application/json
      description: Создать рабочее пространство
      parameters:
      - description: данные нового рабочего пространства
        in: body
        name: newWorkspaceInfo
        required: true
        schema:
          $ref: '#/definitions/dto.NewWorkspaceInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь и его рабочие пространства
          schema:
            $ref: '#/definitions/doc_structs.AllWorkspacesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Создать рабочее пространство
      tags:
      - workspaces
  /workspace/delete/:
    delete:
      consumes:
      - application/json
      description: Удалить рабочее пространство
      parameters:
      - description: id рабочего пространства
        in: body
        name: workspaceID
        required: true
        schema:
          $ref: '#/definitions/dto.WorkspaceID'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Удалить рабочее пространство
      tags:
      - workspaces
  /workspace/update/:
    post:
      consumes:
      - application/json
      description: Обновить рабочее пространство
      parameters:
      - description: обновленные данные рабочего пространства
        in: body
        name: workspaceInfo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatedWorkspaceInfo'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Обновить рабочее пространство
      tags:
      - workspaces
  /workspace/update/change_users/:
    post:
      consumes:
      - application/json
      description: Обновить гостей рабочего пространства
      parameters:
      - description: обновленный список пользователей
        in: body
        name: guestsInfo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatedWorkspaceInfo'
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Обновить гостей рабочего пространства
      tags:
      - workspaces
swagger: "2.0"